Moving core of TKRouter to TripKitAPI
