+++ Sources/TripKit/helpers/TKParserHelper.swift
@@ -19,14 +19,6 @@ enum TKParserHelper {
     return TKRoutingQuery<Never>.parseDate(object)
   }
   
-  static func requestString(for location: TKAPI.Location, includeAddress: Bool = true) -> String {
-    guard includeAddress, let address = location.address else {
-      return String(format: "(%f,%f)", location.latitude, location.longitude)
-    }
-    
-    return String(format: "(%f,%f)\"%@\"", location.latitude, location.longitude, address)
-  }
-  
 #if canImport(MapKit)
   static func requestString(for coordinate: CLLocationCoordinate2D) -> String {
     return String(format: "(%f,%f)", coordinate.latitude, coordinate.longitude)
