+++ Sources/TripKit/model/TKTripAndSegments.swift
@@ -9,19 +9,6 @@
 import Foundation
 
 @objc
-public enum TKTripCostType : Int, Codable {
-  case score
-  case time
-  case duration
-  case price
-  case carbon
-  case hassle
-  case walking
-  case calories
-  case count
-}
-
-@objc
 public enum TKTripSegmentVisibility : Int, Codable {
   
   /// never visible in UI
