+++ Sources/TripKit/server/TKBuzzInfoProvider.swift
@@ -184,7 +184,7 @@ public enum TKBuzzInfoProvider {
   public static func fetchTransitAlerts(for region: TKRegion) async throws -> [TKAPI.AlertMapping] {
     let paras: [String: Any] = [
       "region": region.code,
-      "v": TKSettings.parserJsonVersion
+      "v": TKAPIConfig.parserJsonVersion
     ]
 
     return try await TKServer.shared.hit(
