+++ Sources/TripKitAPI/model/TKTripCostType.swift
@@ -1,0 +1,19 @@
+//
+//  TKTripCostType.swift
+//  TripKit
+//
+//  Created by Adrian Schönig on 20/11/2024.
+//
+
+
+public enum TKTripCostType : Int, Codable {
+  case score
+  case time
+  case duration
+  case price
+  case carbon
+  case hassle
+  case walking
+  case calories
+  case count
+}
